Static image display
This Arduino script is capable of displaying a static image.
The image can be modifyed by re-flashing the Arduino, and swapped by the generated output of the NeoPixel_Matrix_Canvas.py, when exporting the image to .txt.
Initial contained picture within the script is a scenry of Mt. Fuji, with a Japanese pagoda at the right (drawn by me for testing purpose).