3d models, Rev1
These are the .stl files required for the matrix frame. (Rev 1)
Feet.stl: prop-up feet for the frame to be able to stand on a desktop (requires drilling a hole in the frame, and in the feet, and connecting these together with a string, on the bottom).
LED_positions.stl: template for checking the LEDs' positions inside the grid
NeoPixel_Matrix_frame_bottom_frame.stl: bottom half of the frame, enclosing an Arduino Uno on meltable pegs, with proper USB cutout, SD (not micro SD) cutout, and a cutout for the DC 5V barrel jack. The three cutouts on the bottom are for the solder point on the matrix PCB's back, the two little cubes are for holding the feet. The cutouts are labeled.
NeoPixel_Matrix_frame_top_frame_with_grid.stl: top half of the frame. This has a 2mm lip around its perimeter on one side, to make room for the matrix PCB.